blubber: simplify path to HTML doc roots
The current setup doesn't mean HTML content
ends up in /var/www/html like on a default Debian.

We contain everything in /srv/app so this is /srv/app/var/www/html
and we can simplify to /srv/app/html and for consistency
also use html instead of htdocs locally.

Bug: T281538
Change-Id: I6408c43011ffbf26a4716665d715d18b62a9f182