load apache module setenvif
Allows the setting of environment variables based on characteristics of
the request.

This is used in the config of 15.wikipedia.org. An explanation
for why can be found in the comments there.

Since we don't load this by default it leads to
Invalid command 'SetEnvIfNoCase' and the container not starting.

Which in turn lead to Ib0e98a3d6dc6bfa0.

Bug: T300171
Change-Id: Ib9576366ae5e49c02f3c64818a6b5fda813ea607